# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Plagic eSIM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /esim/packages:
    get:
      operationId: listPackages
      tags: [esim]
      summary: List eSIM packages (from esimaccess with markup)
      parameters:
        - name: locationCode
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: pageSize
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Package list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackageListResponse"
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List user orders
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Purchase eSIM package from wallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order with eSIM details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /family/members:
    get:
      operationId: listFamilyMembers
      tags: [family]
      summary: List family members
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Family member list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FamilyMember"
    post:
      operationId: createFamilyMember
      tags: [family]
      summary: Add family member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FamilyMemberInput"
      responses:
        "201":
          description: Created family member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyMember"
  /family/members/{memberId}:
    patch:
      operationId: updateFamilyMember
      tags: [family]
      summary: Update family member
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FamilyMemberUpdate"
      responses:
        "200":
          description: Updated family member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyMember"
    delete:
      operationId: deleteFamilyMember
      tags: [family]
      summary: Remove family member
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /family/members/{memberId}/esims:
    get:
      operationId: listMemberEsims
      tags: [family]
      summary: List eSIMs for a family member
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Member eSIM list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemberEsim"
...
  /redeem/create:
    post:
      operationId: createRedeemCode
      tags: [redeem]
      summary: Purchase and create redeem code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedeemCodeInput"
      responses:
        "201":
          description: Created redeem code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RedeemCode"
  /redeem/apply:
    post:
      operationId: applyRedeemCode
      tags: [redeem]
      summary: Apply redeem code to wallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedeemApplyInput"
      responses:
        "200":
          description: Redeem applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletBalance"
  /redeem/list:
    get:
      operationId: listRedeemCodes
      tags: [redeem]
      summary: List user's redeem codes
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Redeem code list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RedeemCode"
...
components:
  schemas:
...
    EsimPackage:
      type: object
      properties:
        packageCode:
          type: string
        name:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        currency:
          type: string
        dataAmount:
          type: number
        dataUnit:
          type: string
        duration:
          type: integer
        durationUnit:
          type: string
        locationCode:
          type: string
        locationName:
          type: string
        networkType:
          type: string
        type:
          type: string
        slug:
          type: string
        flagEmoji:
          type: ["string", "null"]
      required:
        - packageCode
        - name
        - price
        - currency
        - dataAmount
        - duration
        - locationCode
    PackageListResponse:
      type: object
      properties:
        packages:
          type: array
          items:
            $ref: "#/components/schemas/EsimPackage"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
      required:
        - packages
        - total
...
    WalletBalance:
      type: object
      properties:
        userId:
          type: string
        balance:
          type: number
        currency:
          type: string
      required:
        - userId
        - balance
        - currency
...
    Order:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        familyMemberId:
          type: ["string", "null"]
        packageCode:
          type: string
        packageName:
          type: string
        amount:
          type: number
        status:
          type: string
        iccid:
          type: ["string", "null"]
        qrCode:
          type: ["string", "null"]
        activationCode:
          type: ["string", "null"]
        createdAt:
          type: string
        locationCode:
          type: string
        locationName:
          type: string
        dataAmount:
          type: number
        dataUnit:
          type: string
        duration:
          type: integer
        orderNo:
          type: ["string", "null"]
        wholesaleAmount:
          type: ["number", "null"]
      required:
        - id
        - userId
        - packageCode
        - packageName
        - amount
        - status
        - createdAt
        - locationCode
    OrderInput:
      type: object
      properties:
        userId:
          type: string
        packageCode:
          type: string
        familyMemberId:
          type: ["string", "null"]
      required:
        - userId
        - packageCode
    FamilyMember:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        emoji:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - userId
        - name
        - createdAt
    FamilyMemberInput:
      type: object
      properties:
        userId:
          type: string
        name:
          type: string
        emoji:
          type: ["string", "null"]
      required:
        - userId
        - name
    FamilyMemberUpdate:
      type: object
      properties:
        name:
          type: string
        emoji:
          type: ["string", "null"]
    MemberEsim:
      type: object
      properties:
        orderId:
          type: string
        packageName:
          type: string
        locationCode:
          type: string
        locationName:
          type: string
        iccid:
          type: ["string", "null"]
        status:
          type: string
        dataAmount:
          type: number
        dataUnit:
          type: string
        dataRemaining:
          type: ["number", "null"]
        purchasedAt:
          type: string
      required:
        - orderId
        - packageName
        - locationCode
        - status
        - dataAmount
        - purchasedAt
...
    RedeemCode:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        amount:
          type: number
        currency:
          type: string
        createdBy:
          type: string
        usedBy:
          type: ["string", "null"]
        usedAt:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - code
        - amount
        - currency
        - createdBy
        - status
        - createdAt
    RedeemCodeInput:
      type: object
      properties:
        userId:
          type: string
        amount:
          type: number
      required:
        - userId
        - amount
    RedeemApplyInput:
      type: object
      properties:
        userId:
          type: string
        code:
          type: string
      required:
        - userId
        - code
...