# lib/api-zod/src/generated/types/memberEsim.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Plagic eSIM API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MemberEsim {
  orderId: string;
  packageName: string;
  locationCode: string;
  locationName?: string;
  /** @nullable */
  iccid?: string | null;
  status: string;
  dataAmount: number;
  dataUnit?: string;
  /** @nullable */
  dataRemaining?: number | null;
  purchasedAt: string;
}